requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==5.5.2
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
cloudinary==1.44.1
decorator==5.2.1
future==1.0.0
google-auth==2.40.3
google-genai==1.38.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huepy==1.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
instabot==0.117.0
mock==5.2.0
moviepy==2.2.1
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proglog==0.1.12
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pysocks==1.7.1
pytest==8.4.2
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.25.8
rsa==4.9.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
websockets==15.0.1
